refactor geo_point and geo_node
